#359
- Fixed bug in Welcome Screen
- Removed Hacks Console Module
- Removed fcpx10-3 Module
