--- src/cp/fcpx10-3/init.lua
@@ -1,555 +1,0 @@
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
---                   T H E    M A I N    S C R I P T                          --
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
---------------------------------------------------------------------------------
--- BEGIN MODULE:
---------------------------------------------------------------------------------
-
-local mod = {}
-
---------------------------------------------------------------------------------
--- STANDARD EXTENSIONS:
---------------------------------------------------------------------------------
-
-local application								= require("hs.application")
-local base64									= require("hs.base64")
-local chooser									= require("hs.chooser")
-local console									= require("hs.console")
-local drawing 									= require("hs.drawing")
-local eventtap									= require("hs.eventtap")
-local fnutils 									= require("hs.fnutils")
-local fs										= require("hs.fs")
-local hotkey									= require("hs.hotkey")
-local http										= require("hs.http")
-local image										= require("hs.image")
-local logger									= require("hs.logger")
-local notify									= require("hs.notify")
-local osascript									= require("hs.osascript")
-local pathwatcher								= require("hs.pathwatcher")
-local screen									= require("hs.screen")
-local timer										= require("hs.timer")
-local windowfilter								= require("hs.window.filter")
-
---------------------------------------------------------------------------------
--- EXTERNAL EXTENSIONS:
---------------------------------------------------------------------------------
-
-local ax 										= require("hs._asm.axuielement")
-
---------------------------------------------------------------------------------
--- INTERNAL EXTENSIONS:
---------------------------------------------------------------------------------
-
-local dialog									= require("cp.dialog")
-local fcp										= require("cp.finalcutpro")
-local just										= require("cp.just")
-local metadata									= require("cp.metadata")
-local plist										= require("cp.plist")
-local tools										= require("cp.tools")
-
---------------------------------------------------------------------------------
--- PLUGINS:
---------------------------------------------------------------------------------
-
-local hacksconsole								= require("cp.fcpx10-3.hacksconsole")
-local shortcut									= require("cp.commands.shortcut")
-
---------------------------------------------------------------------------------
--- DEFAULT SETTINGS:
---------------------------------------------------------------------------------
-
-local defaultSettings = {
-												["chooserRememberLast"]							= true,
-												["chooserShowShortcuts"] 						= true,
-												["chooserShowHacks"] 							= true,
-												["chooserShowVideoEffects"] 					= true,
-												["chooserShowAudioEffects"] 					= true,
-												["chooserShowTransitions"] 						= true,
-												["chooserShowTitles"] 							= true,
-												["chooserShowGenerators"] 						= true,
-												["chooserShowMenuItems"]						= true,
-}
-
---------------------------------------------------------------------------------
--- LOCAL VARIABLES:
---------------------------------------------------------------------------------
-
-local execute									= hs.execute
-local log										= logger.new("fcpx10-3")
-
---------------------------------------------------------------------------------
--- MODULE VARIABLES:
---------------------------------------------------------------------------------
-
-mod.releaseColorBoardDown						= false											-- Color Board Shortcut Currently Being Pressed
-mod.finalCutProShortcutKey 						= nil											-- Table of all Final Cut Pro Shortcuts
-mod.finalCutProShortcutKeyPlaceholders 			= nil											-- Table of all needed Final Cut Pro Shortcuts
-mod.lastCommandSet								= nil											-- Last Keyboard Shortcut Command Set
-mod.hacksLoaded 								= false											-- Has FCPX Hacks Loaded Yet?
-mod.isFinalCutProActive 						= false											-- Is Final Cut Pro Active? Used by Watchers.
-
---------------------------------------------------------------------------------
--- RETRIEVES THE PLUGINS MANAGER:
--- If `pluginPath` is provided, the named plugin will be returned. If not,
--- the plugins module is returned.
---------------------------------------------------------------------------------
-function plugins(pluginPath)
-	if not mod._plugins then
-		mod._plugins = require("cp.plugins")
-		mod._plugins.init("cp.plugins")
-	end
-
-	if pluginPath then
-		return mod._plugins(pluginPath)
-	else
-		return mod._plugins
-	end
-end
-
---------------------------------------------------------------------------------
--- RETRIEVES THE MENU MANAGER:
---------------------------------------------------------------------------------
-function menuManager()
-	if not mod._menuManager then
-		mod._menuManager = plugins("cp.plugins.menu.manager")
-
-		--- TODO: Remove this once all menu manaement is migrated to plugins.
-		local manualSection = mod._menuManager.addSection(10000)
-
-	end
-	return mod._menuManager
-end
-
---------------------------------------------------------------------------------
--- LOAD SCRIPT:
---------------------------------------------------------------------------------
-function loadScript()
-
-	--------------------------------------------------------------------------------
-	-- Apply Default Settings:
-	--------------------------------------------------------------------------------
-	for k, v in pairs(defaultSettings) do
-		if metadata.get(k) == nil then
-			metadata.get(k, v)
-		end
-	end
-
-	--------------------------------------------------------------------------------
-	-- Debug Mode:
-	--------------------------------------------------------------------------------
-	log.df("Debug Mode Activated.")
-
-	--------------------------------------------------------------------------------
-	-- Activate Menu Manager
-	--------------------------------------------------------------------------------
-	menuManager()
-
-	--------------------------------------------------------------------------------
-	-- Setup Watches:
-	--------------------------------------------------------------------------------
-
-		--------------------------------------------------------------------------------
-		-- Final Cut Pro Application Watcher:
-		--------------------------------------------------------------------------------
-		fcp:watch({
-			active		= finalCutProActive,
-			inactive	= finalCutProNotActive,
-		})
-
-		--------------------------------------------------------------------------------
-		-- Final Cut Pro Window Watcher:
-		--------------------------------------------------------------------------------
-		finalCutProWindowWatcher()
-
-		--------------------------------------------------------------------------------
-		-- Watch for Final Cut Pro plist Changes:
-		--------------------------------------------------------------------------------
-		preferencesWatcher = pathwatcher.new("~/Library/Preferences/", finalCutProSettingsWatcher):start()
-
-	--------------------------------------------------------------------------------
-	-- Bind Keyboard Shortcuts:
-	--------------------------------------------------------------------------------
-	mod.lastCommandSet = fcp:getActiveCommandSetPath()
-	bindKeyboardShortcuts()
-
-	--------------------------------------------------------------------------------
-	-- Activate the correct modal state:
-	--------------------------------------------------------------------------------
-	if fcp:isFrontmost() then
-		--------------------------------------------------------------------------------
-		-- Used by Watchers to prevent double-ups:
-		--------------------------------------------------------------------------------
-		mod.isFinalCutProActive = true
-
-		--------------------------------------------------------------------------------
-		-- Enable Final Cut Pro Shortcut Keys:
-		--------------------------------------------------------------------------------
-		hotkeys:enter()
-	else
-		--------------------------------------------------------------------------------
-		-- Used by Watchers to prevent double-ups:
-		--------------------------------------------------------------------------------
-		mod.isFinalCutProActive = false
-
-		--------------------------------------------------------------------------------
-		-- Disable Final Cut Pro Shortcut Keys:
-		--------------------------------------------------------------------------------
-		hotkeys:exit()
-	end
-
-	-------------------------------------------------------------------------------
-	-- Set up Chooser:
-	-------------------------------------------------------------------------------
-	hacksconsole.new()
-
-	--------------------------------------------------------------------------------
-	-- All loaded!
-	--------------------------------------------------------------------------------
-	log.df("Successfully loaded.")
-	dialog.displayNotification(metadata.scriptName .. " (v" .. metadata.scriptVersion .. ") " .. i18n("hasLoaded"))
-
-	mod.hacksLoaded = true
-
-end
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
-
-
-
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
---                    K E Y B O A R D     S H O R T C U T S                   --
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
---------------------------------------------------------------------------------
--- DEFAULT SHORTCUT KEYS:
---------------------------------------------------------------------------------
-function defaultShortcutKeys()
-
-	local control					= {"ctrl"}
-	local controlShift 				= {"ctrl", "shift"}
-	local controlOptionCommand 		= {"ctrl", "option", "command"}
-	local controlOptionCommandShift = {"ctrl", "option", "command", "shift"}
-
-    local defaultShortcutKeys = {
-        cpConsole                                             = { characterString = shortcut.textToKeyCode("space"),        modifiers = control,                                fn = function() hacksconsole.show() end,							releasedFn = nil,                                     					repeatFn = nil },
-    }
-	return defaultShortcutKeys
-end
-
---------------------------------------------------------------------------------
--- BIND KEYBOARD SHORTCUTS:
---------------------------------------------------------------------------------
-function bindKeyboardShortcuts()
-
-	--------------------------------------------------------------------------------
-	-- Get Enable Hacks Shortcuts in Final Cut Pro from Settings:
-	--------------------------------------------------------------------------------
-	local enableHacksShortcutsInFinalCutPro = metadata.get("enableHacksShortcutsInFinalCutPro")
-	if enableHacksShortcutsInFinalCutPro == nil then enableHacksShortcutsInFinalCutPro = false end
-
-	--------------------------------------------------------------------------------
-	-- Hacks Shortcuts Enabled:
-	--------------------------------------------------------------------------------
-	if enableHacksShortcutsInFinalCutPro then
-
-		--------------------------------------------------------------------------------
-		-- Get Shortcut Keys from plist:
-		--------------------------------------------------------------------------------
-		mod.finalCutProShortcutKey = nil
-		mod.finalCutProShortcutKey = {}
-		mod.finalCutProShortcutKeyPlaceholders = nil
-		mod.finalCutProShortcutKeyPlaceholders = defaultShortcutKeys()
-
-		--------------------------------------------------------------------------------
-		-- Remove the default shortcut keys:
-		--------------------------------------------------------------------------------
-		for k, v in pairs(mod.finalCutProShortcutKeyPlaceholders) do
-			mod.finalCutProShortcutKeyPlaceholders[k]["characterString"] = ""
-			mod.finalCutProShortcutKeyPlaceholders[k]["modifiers"] = {}
-		end
-
-		--------------------------------------------------------------------------------
-		-- If something goes wrong:
-		--------------------------------------------------------------------------------
-		if getShortcutsFromActiveCommandSet() ~= true then
-			dialog.displayErrorMessage(i18n("customKeyboardShortcutsFailed"))
-			enableHacksShortcutsInFinalCutPro = false
-		end
-
-	end
-
-	--------------------------------------------------------------------------------
-	-- Hacks Shortcuts Disabled:
-	--------------------------------------------------------------------------------
-	if not enableHacksShortcutsInFinalCutPro then
-
-		--------------------------------------------------------------------------------
-		-- Update Active Command Set:
-		--------------------------------------------------------------------------------
-		fcp:getActiveCommandSet(true)
-
-		--------------------------------------------------------------------------------
-		-- Use Default Shortcuts Keys:
-		--------------------------------------------------------------------------------
-		mod.finalCutProShortcutKey = nil
-		mod.finalCutProShortcutKey = defaultShortcutKeys()
-
-	end
-
-	--------------------------------------------------------------------------------
-	-- Reset Modal Hotkey for Final Cut Pro Commands:
-	--------------------------------------------------------------------------------
-	hotkeys = nil
-
-	--------------------------------------------------------------------------------
-	-- Reset Global Hotkeys:
-	--------------------------------------------------------------------------------
-	local currentHotkeys = hotkey.getHotkeys()
-	for _,hk in ipairs(currentHotkeys) do
-		-- only delete hotkeys which are not created by `cp.commands.shortcut`
-		if not hk.shortcut then
-			result = hk:delete()
-		end
-	end
-
-	--------------------------------------------------------------------------------
-	-- Create a modal hotkey object with an absurd triggering hotkey:
-	--------------------------------------------------------------------------------
-	hotkeys = hotkey.modal.new({"command", "shift", "alt", "control"}, "F19")
-
-	--------------------------------------------------------------------------------
-	-- Enable Hotkeys Loop:
-	--------------------------------------------------------------------------------
-	for k, v in pairs(mod.finalCutProShortcutKey) do
-		if v['characterString'] ~= "" and v['fn'] ~= nil then
-			if v['global'] == true then
-				--------------------------------------------------------------------------------
-				-- Global Shortcut:
-				--------------------------------------------------------------------------------
-				hotkey.bind(v['modifiers'], v['characterString'], v['fn'], v['releasedFn'], v['repeatFn'])
-			else
-				--------------------------------------------------------------------------------
-				-- Final Cut Pro Specific Shortcut:
-				--------------------------------------------------------------------------------
-				hotkeys:bind(v['modifiers'], v['characterString'], v['fn'], v['releasedFn'], v['repeatFn'])
-			end
-		end
-	end
-
-	--------------------------------------------------------------------------------
-	-- Enable Hotkeys:
-	--------------------------------------------------------------------------------
-	hotkeys:enter()
-
-	--------------------------------------------------------------------------------
-	-- Let user know that keyboard shortcuts have loaded:
-	--------------------------------------------------------------------------------
-	dialog.displayNotification(i18n("keyboardShortcutsUpdated"))
-
-end
-
---------------------------------------------------------------------------------
--- READ SHORTCUT KEYS FROM FINAL CUT PRO PLIST:
---------------------------------------------------------------------------------
-function getShortcutsFromActiveCommandSet()
-
-	local activeCommandSetTable = fcp:getActiveCommandSet(true)
-
-	if activeCommandSetTable ~= nil then
-		for k, v in pairs(mod.finalCutProShortcutKeyPlaceholders) do
-			local shortcuts = fcp:getCommandShortcuts(k)
-			if shortcuts and #shortcuts > 0 then
-				for x, shortcut in ipairs(shortcuts) do
-
-					local global = v.global or false
-					local xValue = ""
-					if x ~= 1 then xValue = tostring(x) end
-
-					mod.finalCutProShortcutKey[k .. xValue] = {
-						characterString 	= 		shortcut:getKeyCode(),
-						modifiers 			= 		shortcut:getModifiers(),
-						fn 					= 		mod.finalCutProShortcutKeyPlaceholders[k]['fn'],
-						releasedFn 			= 		mod.finalCutProShortcutKeyPlaceholders[k]['releasedFn'],
-						repeatFn 			= 		mod.finalCutProShortcutKeyPlaceholders[k]['repeatFn'],
-						global 				= 		global,
-					}
-				end
-			end
-		end
-		return true
-	else
-		return false
-	end
-
-end
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
-
-
-
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
---                             W A T C H E R S                                --
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
---------------------------------------------------------------------------------
--- AUTOMATICALLY DO THINGS WHEN FINAL CUT PRO IS ACTIVATED OR DEACTIVATED:
---------------------------------------------------------------------------------
-function finalCutProWatcher(appName, eventType, appObject)
-	if (appName == "Final Cut Pro") then
-		if (eventType == application.watcher.activated) then
-			finalCutProActive()
-		elseif (eventType == application.watcher.deactivated) or (eventType == application.watcher.terminated) then
-			finalCutProNotActive()
-		end
-	end
-end
-
---------------------------------------------------------------------------------
--- AUTOMATICALLY DO THINGS WHEN FINAL CUT PRO WINDOWS ARE CHANGED:
---------------------------------------------------------------------------------
-function finalCutProWindowWatcher()
-
-	-- Watch the command editor showing and hiding.
-	fcp:commandEditor():watch({
-		show = function(commandEditor)
-			--------------------------------------------------------------------------------
-			-- Disable Hotkeys:
-			--------------------------------------------------------------------------------
-			if hotkeys ~= nil then -- For the rare case when Command Editor is open on load.
-				log.df("Disabling Hotkeys")
-				hotkeys:exit()
-			end
-		end,
-		hide = function(commandEditor)
-			--------------------------------------------------------------------------------
-			-- Refresh Keyboard Shortcuts:
-			--------------------------------------------------------------------------------
-			timer.doAfter(0.0000000000001, function() bindKeyboardShortcuts() end)
-			--------------------------------------------------------------------------------
-		end
-	})
-end
-
-	--------------------------------------------------------------------------------
-	-- Final Cut Pro Active:
-	--------------------------------------------------------------------------------
-	function finalCutProActive()
-		--------------------------------------------------------------------------------
-		-- Only do once:
-		--------------------------------------------------------------------------------
-		if mod.isFinalCutProActive then return end
-		mod.isFinalCutProActive = true
-
-		--------------------------------------------------------------------------------
-		-- Don't trigger until after the script has loaded:
-		--------------------------------------------------------------------------------
-		if not mod.hacksLoaded then
-			timer.waitUntil(function() return mod.hacksLoaded end, function()
-				if fcp:isFrontmost() then
-					mod.isFinalCutProActive = false
-					finalCutProActive()
-				end
-			end, 0.1)
-			return
-		end
-
-		--------------------------------------------------------------------------------
-		-- Enable Hotkeys:
-		--------------------------------------------------------------------------------
-		timer.doAfter(0.0000000000001, function()
-			hotkeys:enter()
-		end)
-
-		--------------------------------------------------------------------------------
-		-- Update Current Language:
-		--------------------------------------------------------------------------------
-		timer.doAfter(0.0000000000001, function()
-			fcp:getCurrentLanguage(true)
-		end)
-
-	end
-
-	--------------------------------------------------------------------------------
-	-- Final Cut Pro Not Active:
-	--------------------------------------------------------------------------------
-	function finalCutProNotActive()
-		--------------------------------------------------------------------------------
-		-- Only do once:
-		--------------------------------------------------------------------------------
-		if not mod.isFinalCutProActive then return end
-		mod.isFinalCutProActive = false
-
-		--------------------------------------------------------------------------------
-		-- Don't trigger until after the script has loaded:
-		--------------------------------------------------------------------------------
-		if not mod.hacksLoaded then return end
-
-		--------------------------------------------------------------------------------
-		-- Disable hotkeys:
-		--------------------------------------------------------------------------------
-		hotkeys:exit()
-
-		--------------------------------------------------------------------------------
-		-- Delete the Mouse Circle:
-		--------------------------------------------------------------------------------
-		plugins("cp.plugins.browser.playhead").deleteHighlight()
-	end
-
---------------------------------------------------------------------------------
--- AUTOMATICALLY DO THINGS WHEN FCPX PLIST IS UPDATED:
---------------------------------------------------------------------------------
-function finalCutProSettingsWatcher(files)
-    doReload = false
-    for _,file in pairs(files) do
-        if file:sub(-24) == "com.apple.FinalCut.plist" then
-            doReload = true
-        end
-    end
-    if doReload then
-
-		--------------------------------------------------------------------------------
-		-- Refresh Keyboard Shortcuts if Command Set Changed & Command Editor Closed:
-		--------------------------------------------------------------------------------
-    	if mod.lastCommandSet ~= fcp:getActiveCommandSetPath() then
-    		if not fcp:commandEditor():isShowing() then
-	    		timer.doAfter(0.0000000000001, function() bindKeyboardShortcuts() end)
-			end
-		end
-
-    	--------------------------------------------------------------------------------
-    	-- Update Menubar Icon:
-    	--------------------------------------------------------------------------------
-    	timer.doAfter(0.0000000000001, function() menuManager():updateMenubarIcon() end)
-    end
-end
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
-
-
-
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
---                L E T ' S     D O     T H I S     T H I N G !               --
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
-
-loadScript()
-
---------------------------------------------------------------------------------
---------------------------------------------------------------------------------
